import: Move 'process_uploads' and 'process_emojis' to import_util.
